requirements.txt
Django~=5.1
djangorestframework~=3.15.2
djangorestframework-simplejwt~=5.3.1
drf-spectacular==0.27.2
django-cors-headers==4.4.0
django-rest-passwordreset==1.4.1
